add Tsp example

+++ src/main/java/com/github/bibenga/alns/State.java
@@ -4,5 +4,5 @@ public interface State {
     /*
     Computes the state's associated objective value.
     */
-    float objective();
+    double objective();
 }
